user_testing/analysis_software/pyproject.toml
[project]
name = "user-testing-analysis"
version = "0.1.0"
description = "User testing analysis toolkit starting with video transcription"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0,<2.5.0",
    "transformers>=4.30.0",
    "ffmpeg-python>=0.2.0",
    "tqdm>=4.65.0",
    "mistral-common>=1.8.3",
    "librosa>=0.11.0",
    "soundfile>=0.13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "isort>=6.0.1",
]
